refactor CustomLabelGenerator
